Separar testes por arquivo e adicionar inserção de usuario e senha criptografados no config.ini
