+++ README.md
@@ -66,3 +66,6 @@ Para verificar se o seu código encontra-se bem indentado e bonito, use o seguin
 make style
 ```
 Importante usar este comando antes de mandar uma PR para garantir código mais legível.
+
+# >>> plain_text = cipher_suite.decrypt(cipher_text)
+# >>> print(str(plain_text,'utf-8'))
